requirements.txt
aiosmtplib==3.0.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.2
binaryornot==0.4.4
blinker==1.9.0
boto3==1.40.8
botocore==1.40.8
bson==0.5.10
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.9.1
cryptography==44.0.2
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.1.7
django-browser-reload==1.18.0
django-guardian==3.2.0
django-pydantic-field==0.3.12
django-redis==5.4.0
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.4.2
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
numpy==2.3.2
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.13.2
pyobjectid==0.1.4
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
rsa==4.9
s3transfer==0.13.1
setuptools==76.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
